requirements-dev.txt
# Development and testing dependencies
# Install with: pip install -r requirements-dev.txt

# Testing
factory_boy>=3.3.0,<4.0
pytest>=8.0.0,<9.0
pytest-django>=4.9.0,<5.0
pytest-xdist>=3.5.0,<4.0
pytest-cov>=6.0.0,<7.0
pytest-factoryboy>=2.7.0,<3.0
freezegun>=1.4.0,<2.0
faker>=33.0.0,<40.0

# Docker and container testing
docker>=7.0.0,<8.0
pyyaml>=6.0,<7.0
requests>=2.28.0,<3.0

# Linting and code quality
ruff>=0.8.0,<1.0

# Security scanning
bandit>=1.9.0,<2.0
pip-audit>=2.9.0,<3.0

# Development utilities
django-debug-toolbar>=4.2.0,<7.0

# Note: These are development-only dependencies
# Production requirements are in requirements.txt
